app/core/config.py
import os
from dotenv import load_dotenv

load_dotenv()

SWAPI_BASE_URL = os.getenv("SWAPI_BASE_URL")
DATABASE_URL = os.getenv("DATABASE_URL")
SECRET_KEY = os.getenv("SECRET_KEY")
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = int(os.getenv("ACCESS_TOKEN_EXPIRE_MINUTES", "30"))