requirements.txt
cachetools==6.2.6
fastapi==0.128.0
httpx==0.28.1
passlib==1.7.4
pydantic==2.12.5
pytest==9.0.2
python-dotenv==1.2.1
python_jose==3.5.0
SQLAlchemy==2.0.46
